fix all warnings on Win32 for x86 and x64 builds

+++ connection.cpp
@@ -2,7 +2,7 @@
 
 using namespace rrl;
 
-Connection::Connection(int sockfd)
+Connection::Connection(intptr_t sockfd)
     : socket_(sockfd)
 {}
